api/requirements.txt
azure-functions
google-generativeai
python-dotenv 